Added Cypres Priettier & Redux Toolkit to project
