+++ .pretierIgnore
@@ -1,0 +1,3 @@
+/node_modules
+/.pnp
+.pnp.js
